+++ flutter_inappwebview_platform_interface/lib/src/types/main.dart
@@ -237,3 +237,4 @@ export 'process_failed_reason.dart' show ProcessFailedReason;
 export 'process_failed_detail.dart' show ProcessFailedDetail;
 export 'focus_direction.dart' show FocusDirection;
 export 'enum_method.dart';
+export 'pdf_toolbar_items.dart' show PdfToolbarItems;